* Add relations between one user and many decks  * Move creating deck to user resource by POST /users/{id}/decks  * Add minor exception handling with controller advice

+++ src/main/java/com/parab0la/gotedh/repository/DeckRepository.java
@@ -3,5 +3,5 @@ package com.parab0la.gotedh.repository;
 import com.parab0la.gotedh.model.Deck;
 import org.springframework.data.repository.CrudRepository;
 
-public interface DeckRepository extends CrudRepository<Deck, Integer> {
+public interface DeckRepository extends CrudRepository<Deck, Long> {
 }
